+++ package.json
@@ -5,17 +5,18 @@
   "main": "index.ts",
   "_moduleAliases": {
     "@appTypes": "src/types",
-    "@errors": "src/errors",
-    "@db_modules": "src/db_models",
-    "@mutation": "src/mutations/index.ts",
-    "@queries": "src/queries/index.ts",
-    "@schemas": "src/schemas",
-    "@subscriptions": "src/subscriptions",
+    "@errors": "src/services/errors",
+    "@services": "src/services",
+    "@models": "src/database/models",
+    "@mutation": "src/graphql/mutations/index.ts",
+    "@queries": "src/graphql/queries/index.ts",
+    "@schemas": "src/graphql/schemas/index.ts",
+    "@subscriptions": "src/graphql/subscriptions",
     "@utils": "src/utils",
-    "@appConfig": "app_config.json"
+    "@appConfig": "src/appConfig.ts"
   },
   "scripts": {
-    "start": "nodemon --exec ts-node index.ts"
+    "dev": "nodemon --exec ts-node index.ts"
   },
   "author": "",
   "license": "ISC",
@@ -32,11 +33,13 @@
   },
   "dependencies": {
     "@types/bcrypt": "^3.0.0",
+    "@types/dotenv": "^6.1.1",
     "@types/jsonwebtoken": "^8.3.2",
     "@types/mongoose": "^5.5.9",
     "@types/yup": "^0.26.22",
     "apollo-server": "^2.7.0",
     "bcrypt": "^3.0.6",
+    "dotenv": "^8.1.0",
     "express": "^4.17.1",
     "graphql": "^14.4.2",
     "jsonwebtoken": "^8.5.1",
