+++ src/services/errors/userErrors.ts
@@ -1,14 +1,5 @@
 import { TErrorFuncParams } from "@appTypes/error";
 
-export const HashingErr = ({ from, errItself }: TErrorFuncParams) =>
-  JSON.stringify({
-    statusCode: 400,
-    name: "Hashing Error",
-    msg: "An error occured while hashing the password",
-    from,
-    errItself
-  });
-
 export const ExistingEmail = ({ from, errItself }: TErrorFuncParams) =>
   JSON.stringify({
     statusCode: 400,
